Revert "Feat/increase fee decimal precision (#282)"
This reverts commit 42f93624a21243f06d59bd7bc8ad8ee6a0ce0903.

+++ test/invariant/InvariantTestBase.t.sol
@@ -244,7 +244,7 @@ abstract contract InvariantTestBase is Test {
         leverageManager.setTreasuryActionFee(action, newTreasuryFee);
     }
 
-    function _setManagementFee(ILeverageToken leverageToken, uint128 newManagementFee) internal {
+    function _setManagementFee(ILeverageToken leverageToken, uint256 newManagementFee) internal {
         vm.prank(feeManagerRole);
         leverageManager.setManagementFee(leverageToken, newManagementFee);
     }
